Extract setup logic
[ghstack-poisoned]
