+++ apeiron/utils.py
@@ -1,0 +1,8 @@
+def parse_feature_gates(feature_gates_str: str) -> dict[str, bool]:
+    """Parse feature gates from a string into a dictionary."""
+    feature_gates_dict = {}
+    for feature_gate in feature_gates_str.split(","):
+        feature_gate = feature_gate.strip()
+        if feature_gate:
+            feature_gates_dict[feature_gate] = True
+    return feature_gates_dict
